AbyStore refactoring and module documentation
