+++ iris-mpc-cpu/benches/hnsw.rs
@@ -4,7 +4,7 @@ use iris_mpc_common::iris_db::{db::IrisDB, iris::IrisCode};
 use iris_mpc_cpu::{
     database_generators::{create_random_sharing, generate_galois_iris_shares},
     execution::local::LocalRuntime,
-    hawkers::{aby3_store::Aby3Store, plaintext_store::PlaintextStore},
+    hawkers::{aby3::test_utils::lazy_setup_from_files_with_grpc, plaintext_store::PlaintextStore},
     hnsw::{GraphMem, HnswSearcher},
     protocol::ops::{
         batch_signed_lift_vec, cross_compare, galois_ring_pairwise_distance, galois_ring_to_rep3,
@@ -176,7 +176,7 @@ fn bench_gr_ready_made_hnsw(c: &mut Criterion) {
 
         let secret_searcher = rt.block_on(async move {
             let mut rng = AesRng::seed_from_u64(0_u64);
-            Aby3Store::lazy_setup_from_files_with_grpc(
+            lazy_setup_from_files_with_grpc(
                 "./data/store.ndjson",
                 &format!("./data/graph_{}.dat", database_size),
                 &mut rng,
