+++ iris-mpc-cpu/src/hawkers/aby3/aby3_store.rs
@@ -1,47 +1,34 @@
 use crate::{
-    database_generators::{generate_galois_iris_shares, GaloisRingSharedIris},
+    database_generators::GaloisRingSharedIris,
     execution::{
-        local::{generate_local_identities, LocalRuntime},
         player::Identity,
         session::{Session, SessionHandles},
     },
-    hawkers::plaintext_store::{PlaintextStore, PointId},
-    hnsw::{
-        graph::layered_graph::{GraphMem, Layer},
-        HnswSearcher, SortedNeighborhood, VectorStore,
-    },
-    network::NetworkType,
+    hawkers::plaintext_store::PointId,
+    hnsw::VectorStore,
     protocol::ops::{
         batch_signed_lift_vec, compare_threshold_and_open, cross_compare,
         galois_ring_pairwise_distance, galois_ring_to_rep3,
     },
-    py_bindings::{io::read_bin, plaintext_store::from_ndjson_file},
     shares::{
         ring_impl::RingElement,
         share::{DistanceShare, Share},
     },
 };
-use aes_prng::AesRng;
-use iris_mpc_common::iris_db::db::IrisDB;
-use rand::{CryptoRng, RngCore, SeedableRng};
 use serde::{Deserialize, Serialize};
 use std::{
-    collections::HashMap,
     fmt::{Debug, Display},
     num::ParseIntError,
     str::FromStr,
     sync::Arc,
     vec,
 };
-use tokio::{
-    sync::{RwLock, RwLockWriteGuard},
-    task::JoinSet,
-};
+use tokio::sync::{RwLock, RwLockWriteGuard};
 use tracing::instrument;
 
 #[derive(Copy, Default, Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub struct VectorId {
-    id: PointId,
+    pub(crate) id: PointId,
 }
 
 impl Display for VectorId {
@@ -167,49 +154,6 @@ impl SharedIrisesRef {
     }
 }
 
-pub fn setup_local_player_preloaded_db(
-    database: Vec<GaloisRingSharedIris>,
-) -> eyre::Result<SharedIrisesRef> {
-    let aby3_store = SharedIrisesRef::new(database);
-    Ok(aby3_store)
-}
-
-pub async fn setup_local_aby3_players_with_preloaded_db<R: RngCore + CryptoRng>(
-    rng: &mut R,
-    plain_store: &PlaintextStore,
-    network_t: NetworkType,
-) -> eyre::Result<Vec<Aby3Store>> {
-    let identities = generate_local_identities();
-
-    let mut shared_irises = vec![vec![]; identities.len()];
-
-    for iris in plain_store.points.iter() {
-        let all_shares = generate_galois_iris_shares(rng, iris.data.0.clone());
-        for (i, shares) in all_shares.iter().enumerate() {
-            shared_irises[i].push(shares.clone());
-        }
-    }
-
-    let storages: Vec<SharedIrisesRef> = shared_irises
-        .into_iter()
-        .map(|player_irises| setup_local_player_preloaded_db(player_irises).unwrap())
-        .collect();
-    let runtime = LocalRuntime::mock_setup(network_t).await?;
-
-    identities
-        .into_iter()
-        .zip(storages.into_iter())
-        .map(|(identity, storage)| {
-            let session = runtime.get_session(&identity)?;
-            Ok(Aby3Store {
-                session,
-                storage,
-                owner: identity,
-            })
-        })
-        .collect()
-}
-
 /// Implementation of VectorStore based on the ABY3 framework (<https://eprint.iacr.org/2018/403.pdf>).
 ///
 /// Note that all SMPC operations are performed in a single session.
@@ -224,23 +168,6 @@ impl Aby3Store {
     pub fn get_owner_index(&self) -> usize {
         self.session.boot_session.own_role().unwrap().index()
     }
-}
-
-pub async fn setup_local_store_aby3_players(
-    network_t: NetworkType,
-) -> eyre::Result<Vec<Aby3Store>> {
-    let runtime = LocalRuntime::mock_setup(network_t).await?;
-    generate_local_identities()
-        .into_iter()
-        .map(|identity| {
-            let session = runtime.get_session(&identity)?;
-            Ok(Aby3Store {
-                session,
-                storage: SharedIrisesRef::default(),
-                owner: identity,
-            })
-        })
-        .collect()
 }
 
 impl Aby3Store {
@@ -367,7 +294,7 @@ impl Aby3Store {
     }
 
     #[instrument(level = "trace", target = "searcher::network", skip_all)]
-    async fn eval_distance_vectors(
+    pub(crate) async fn eval_distance_vectors(
         &mut self,
         vector1: &<Aby3Store as VectorStore>::VectorRef,
         vector2: &<Aby3Store as VectorStore>::VectorRef,
@@ -382,282 +309,27 @@ impl Aby3Store {
     }
 }
 
-impl Aby3Store {
-    /// Converts a plaintext graph store to a secret-shared graph store.
-    ///
-    /// If recompute_distance is true, distances are recomputed from scratch via
-    /// SMPC. Otherwise, distances are naively converted from plaintext ones
-    /// via trivial shares,
-    /// i.e., the sharing of a value x is a triple (x, 0, 0).
-    async fn graph_from_plain(
-        &mut self,
-        graph_store: &GraphMem<PlaintextStore>,
-        recompute_distances: bool,
-    ) -> GraphMem<Aby3Store> {
-        let ep = graph_store.get_entry_point().await;
-        let new_ep = ep.map(|(vector_ref, layer_count)| (VectorId { id: vector_ref }, layer_count));
-
-        let layers = graph_store.get_layers();
-
-        let mut shared_layers = vec![];
-        for layer in layers {
-            let links = layer.get_links_map();
-            let mut shared_links = HashMap::new();
-            for (source_v, queue) in links {
-                let source_v = source_v.into();
-                let mut shared_queue = vec![];
-                for (target_v, dist) in queue.as_vec_ref() {
-                    let target_v = target_v.into();
-                    let distance = if recompute_distances {
-                        // recompute distances of graph edges from scratch
-                        self.eval_distance_vectors(&source_v, &target_v).await
-                    } else {
-                        // convert plaintext distances to trivial shares, i.e., d -> (d, 0, 0)
-                        DistanceShare::new(
-                            self.get_trivial_share(dist.0),
-                            self.get_trivial_share(dist.1),
-                        )
-                    };
-                    shared_queue.push((target_v, distance.clone()));
-                }
-                shared_links.insert(
-                    source_v,
-                    SortedNeighborhood::from_ascending_vec(shared_queue.clone()),
-                );
-            }
-            shared_layers.push(Layer::from_links(shared_links));
-        }
-        GraphMem::from_precomputed(new_ep, shared_layers)
-    }
-
-    /// Generates 3 pairs of vector stores and graphs from a plaintext
-    /// vector store and graph read from disk, which are returned as well.
-    /// The network type is specified by the user.
-    /// A recompute flag is used to determine whether to recompute the distances
-    /// from stored shares. If recompute is set to false, the distances are
-    /// naively converted from plaintext.
-    pub async fn lazy_setup_from_files<R: RngCore + Clone + CryptoRng>(
-        plainstore_file: &str,
-        plaingraph_file: &str,
-        rng: &mut R,
-        database_size: usize,
-        network_t: NetworkType,
-        recompute_distances: bool,
-    ) -> eyre::Result<(
-        (PlaintextStore, GraphMem<PlaintextStore>),
-        Vec<(Self, GraphMem<Self>)>,
-    )> {
-        if database_size > 100_000 {
-            return Err(eyre::eyre!("Database size too large, max. 100,000"));
-        }
-        let generation_comment = "Please, generate benchmark data with cargo run --release --bin \
-                                  generate_benchmark_data.";
-        let plaintext_vector_store = from_ndjson_file(plainstore_file, Some(database_size))
-            .map_err(|e| eyre::eyre!("Cannot find store: {e}. {generation_comment}"))?;
-        let plaintext_graph_store: GraphMem<PlaintextStore> = read_bin(plaingraph_file)
-            .map_err(|e| eyre::eyre!("Cannot find graph: {e}. {generation_comment}"))?;
-
-        let protocol_stores =
-            setup_local_aby3_players_with_preloaded_db(rng, &plaintext_vector_store, network_t)
-                .await?;
-
-        let mut jobs = JoinSet::new();
-        for store in protocol_stores.iter() {
-            let mut store = store.clone();
-            let plaintext_graph_store = plaintext_graph_store.clone();
-            jobs.spawn(async move {
-                (
-                    store.clone(),
-                    store
-                        .graph_from_plain(&plaintext_graph_store, recompute_distances)
-                        .await,
-                )
-            });
-        }
-        let mut secret_shared_stores = jobs.join_all().await;
-        secret_shared_stores.sort_by_key(|(store, _)| store.get_owner_index());
-        let plaintext = (plaintext_vector_store, plaintext_graph_store);
-        Ok((plaintext, secret_shared_stores))
-    }
-
-    /// Generates 3 pairs of vector stores and graphs from a plaintext
-    /// vector store and graph read from disk, which are returned as well.
-    /// Networking is based on gRPC.
-    pub async fn lazy_setup_from_files_with_grpc<R: RngCore + Clone + CryptoRng>(
-        plainstore_file: &str,
-        plaingraph_file: &str,
-        rng: &mut R,
-        database_size: usize,
-        recompute_distances: bool,
-    ) -> eyre::Result<(
-        (PlaintextStore, GraphMem<PlaintextStore>),
-        Vec<(Self, GraphMem<Self>)>,
-    )> {
-        Self::lazy_setup_from_files(
-            plainstore_file,
-            plaingraph_file,
-            rng,
-            database_size,
-            NetworkType::GrpcChannel,
-            recompute_distances,
-        )
-        .await
-    }
-
-    /// Generates 3 pairs of vector stores and graphs from a random plaintext
-    /// vector store and graph, which are returned as well.
-    /// The network type is specified by the user.
-    /// A recompute flag is used to determine whether to recompute the distances
-    /// from stored shares. If recompute is set to false, the distances are
-    /// naively converted from plaintext.
-    pub async fn lazy_random_setup<R: RngCore + Clone + CryptoRng>(
-        rng: &mut R,
-        database_size: usize,
-        network_t: NetworkType,
-        recompute_distances: bool,
-    ) -> eyre::Result<(
-        (PlaintextStore, GraphMem<PlaintextStore>),
-        Vec<(Self, GraphMem<Self>)>,
-    )> {
-        let searcher = HnswSearcher::default();
-        let (plaintext_vector_store, plaintext_graph_store) =
-            PlaintextStore::create_random(rng, database_size, &searcher).await?;
-
-        let protocol_stores =
-            setup_local_aby3_players_with_preloaded_db(rng, &plaintext_vector_store, network_t)
-                .await?;
-
-        let mut jobs = JoinSet::new();
-        for store in protocol_stores.iter() {
-            let mut store = store.clone();
-            let plaintext_graph_store = plaintext_graph_store.clone();
-            jobs.spawn(async move {
-                (
-                    store.clone(),
-                    store
-                        .graph_from_plain(&plaintext_graph_store, recompute_distances)
-                        .await,
-                )
-            });
-        }
-        let mut secret_shared_stores = jobs.join_all().await;
-        secret_shared_stores.sort_by_key(|(store, _)| store.get_owner_index());
-        let plaintext = (plaintext_vector_store, plaintext_graph_store);
-        Ok((plaintext, secret_shared_stores))
-    }
-
-    /// Generates 3 pairs of vector stores and graphs from a random plaintext
-    /// vector store and graph, which are returned as well. Networking is
-    /// based on local async_channel.
-    pub async fn lazy_random_setup_with_local_channel<R: RngCore + Clone + CryptoRng>(
-        rng: &mut R,
-        database_size: usize,
-        recompute_distances: bool,
-    ) -> eyre::Result<(
-        (PlaintextStore, GraphMem<PlaintextStore>),
-        Vec<(Aby3Store, GraphMem<Aby3Store>)>,
-    )> {
-        Self::lazy_random_setup(
-            rng,
-            database_size,
-            NetworkType::LocalChannel,
-            recompute_distances,
-        )
-        .await
-    }
-
-    /// Generates 3 pairs of vector stores and graphs from a random plaintext
-    /// vector store and graph, which are returned as well. Networking is
-    /// based on gRPC.
-    pub async fn lazy_random_setup_with_grpc<R: RngCore + Clone + CryptoRng>(
-        rng: &mut R,
-        database_size: usize,
-        recompute_distances: bool,
-    ) -> eyre::Result<(
-        (PlaintextStore, GraphMem<PlaintextStore>),
-        Vec<(Aby3Store, GraphMem<Aby3Store>)>,
-    )> {
-        Self::lazy_random_setup(
-            rng,
-            database_size,
-            NetworkType::GrpcChannel,
-            recompute_distances,
-        )
-        .await
-    }
-
-    /// Generates 3 pairs of vector stores and graphs corresponding to each
-    /// local player.
-    pub async fn shared_random_setup<R: RngCore + Clone + CryptoRng>(
-        rng: &mut R,
-        database_size: usize,
-        network_t: NetworkType,
-    ) -> eyre::Result<Vec<(Self, GraphMem<Self>)>> {
-        let rng_searcher = AesRng::from_rng(rng.clone())?;
-        let cleartext_database = IrisDB::new_random_rng(database_size, rng).db;
-        let shared_irises: Vec<_> = (0..database_size)
-            .map(|id| generate_galois_iris_shares(rng, cleartext_database[id].clone()))
-            .collect();
-
-        let mut local_stores = setup_local_store_aby3_players(network_t).await?;
-
-        let mut jobs = JoinSet::new();
-        for store in local_stores.iter_mut() {
-            let mut store = store.clone();
-            let role = store.get_owner_index();
-            let mut rng_searcher = rng_searcher.clone();
-            let queries = (0..database_size)
-                .map(|id| store.prepare_query(shared_irises[id][role].clone()))
-                .collect::<Vec<_>>();
-            jobs.spawn(async move {
-                let mut graph_store = GraphMem::new();
-                let searcher = HnswSearcher::default();
-                // insert queries
-                for query in queries.iter() {
-                    searcher
-                        .insert(&mut store, &mut graph_store, query, &mut rng_searcher)
-                        .await;
-                }
-                (store, graph_store)
-            });
-        }
-        let mut result = jobs.join_all().await;
-        // preserve order of players
-        result.sort_by(|(store1, _), (store2, _)| {
-            store1.get_owner_index().cmp(&store2.get_owner_index())
-        });
-        Ok(result)
-    }
-
-    /// Generates 3 pairs of vector stores and graphs corresponding to each
-    /// local player. Networking is based on local async_channel.
-    pub async fn shared_random_setup_with_local_channel<R: RngCore + Clone + CryptoRng>(
-        rng: &mut R,
-        database_size: usize,
-    ) -> eyre::Result<Vec<(Self, GraphMem<Self>)>> {
-        Self::shared_random_setup(rng, database_size, NetworkType::LocalChannel).await
-    }
-
-    /// Generates 3 pairs of vector stores and graphs corresponding to each
-    /// local player. Networking is based on gRPC.
-    pub async fn shared_random_setup_with_grpc<R: RngCore + Clone + CryptoRng>(
-        rng: &mut R,
-        database_size: usize,
-    ) -> eyre::Result<Vec<(Self, GraphMem<Self>)>> {
-        Self::shared_random_setup(rng, database_size, NetworkType::GrpcChannel).await
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use std::collections::HashMap;
+
     use super::*;
     use crate::{
         database_generators::generate_galois_iris_shares,
+        hawkers::{
+            aby3::test_utils::{
+                lazy_random_setup, setup_local_store_aby3_players, shared_random_setup,
+            },
+            plaintext_store::PlaintextStore,
+        },
         hnsw::{GraphMem, HnswSearcher},
+        network::NetworkType,
     };
     use aes_prng::AesRng;
+    use iris_mpc_common::iris_db::db::IrisDB;
     use itertools::Itertools;
     use rand::SeedableRng;
+    use tokio::task::JoinSet;
     use tracing_test::traced_test;
 
     #[tokio::test(flavor = "multi_thread")]
@@ -725,15 +397,14 @@ mod tests {
         let database_size = 10;
         let network_t = NetworkType::LocalChannel;
         let (mut cleartext_data, secret_data) =
-            Aby3Store::lazy_random_setup(&mut rng, database_size, network_t.clone(), true)
+            lazy_random_setup(&mut rng, database_size, network_t.clone(), true)
                 .await
                 .unwrap();
 
         let mut rng = AesRng::seed_from_u64(0_u64);
-        let vector_graph_stores =
-            Aby3Store::shared_random_setup(&mut rng, database_size, network_t)
-                .await
-                .unwrap();
+        let vector_graph_stores = shared_random_setup(&mut rng, database_size, network_t)
+            .await
+            .unwrap();
 
         for ((v_from_scratch, _), (premade_v, _)) in
             vector_graph_stores.iter().zip(secret_data.iter())
@@ -892,7 +563,7 @@ mod tests {
         let database_size = 2;
         let searcher = HnswSearcher::default();
         let mut vectors_and_graphs =
-            Aby3Store::shared_random_setup(&mut rng, database_size, NetworkType::LocalChannel)
+            shared_random_setup(&mut rng, database_size, NetworkType::LocalChannel)
                 .await
                 .unwrap();
