+++ iris-mpc-cpu/src/hawkers/aby3/mod.rs
@@ -1,0 +1,6 @@
+/// Store with vectors in secret shared form.
+/// The underlying operations are secure multi-party computation (MPC) operations.
+pub mod aby3_store;
+
+/// Various utilities for data generation in tests.
+pub mod test_utils;
